Enable jackknife+ predictive intervals for a fit model

Prepares for jackknife+ conformal prediction by performing Pareto-smoothed importance sampling to yield leave-one-out residuals.

Usage

loo_conformal(fit, ...)

# S3 method for default
loo_conformal(fit, truth, chain = NULL, est_fun = c("mean", "median"), ...)

# S3 method for stanreg
loo_conformal(fit, est_fun = c("mean", "median"), ...)

# S3 method for brmsfit
loo_conformal(fit, est_fun = c("mean", "median"), ...)

# S3 method for array
loo_conformal(
  fit,
  log_lik,
  truth,
  chain = NULL,
  est_fun = c("mean", "median"),
  ...
)

Arguments

fit

Model fit; an object with posterior_predict() and log_lik() methods. Can also be an array of posterior predictions.

...

Ignored.

truth

True values to predict. Not required for rstanarm or brms models.

chain

An integer vector identifying the chain numbers for the posterior draws. Should be provided if multiple chains are used.

est_fun

Whether to use the posterior mean (the default) or median as a point estimate.

log_lik

If fit is an array, then log_lik should be an array of matching dimension giving the pointwise log-likelihood.

Value

A modified fit object with an additional class conformal. Calling predictive_interval() on this new object will yield conformal intervals.

References

Vehtari, A., Simpson, D., Gelman, A., Yao, Y., & Gabry, J. (2015). Pareto smoothed importance sampling. arXiv preprint arXiv:1507.02646.

Examples