Jackknife(+) predictive intervals

Construct finite-sample calibrated predictive intervals for Bayesian models, following the approach in Barber et al. (2021). By default, the intervals will also reflect the relative uncertainty in the Bayesian model, using the locally-weighted conformal methods of Lei et al. (2018).

Usage

# S3 method for conformal
predictive_interval(object, probs = 0.9, plus = NULL, local = TRUE, ...)

Arguments

object

A fitted model which has been passed through loo_conformal()

probs

The coverage probabilities to calculate intervals for. Empirically, the coverage rate of the constructed intervals will generally match these probabilities, but the theoretical guarantee for a probability of \(1-\alpha\) is only for coverage of at least \(1-2\alpha\), and only if plus=TRUE (below).

plus

If TRUE, construct jackknife+ intervals, which have a theoretical guarantee. These require higher computational costs, which scale with both the number of training and prediction points. Defaults to TRUE when both of these numbers are less than 500.

local

If TRUE (the default), perform locally-weighted conformal inference. This will inflate the width of the predictive intervals by a constant amount across all predictions, preserving the relative amount of uncertainty captured by the model. If FALSE, all predictive intervals will have (nearly) the same width.

...

Further arguments to the posterior_predict() method for object.

Value

A matrix with the number of rows matching the number of predictions. Columns will be labeled with a percentile corresponding to probs; e.g. if probs=0.9 the columns will be 5% and 95%.

References

Barber, R. F., Candes, E. J., Ramdas, A., & Tibshirani, R. J. (2021). Predictive inference with the jackknife+. The Annals of Statistics, 49(1), 486-507.

Lei, J., G’Sell, M., Rinaldo, A., Tibshirani, R. J., & Wasserman, L. (2018). Distribution-free predictive inference for regression. Journal of the American Statistical Association, 113(523), 1094-1111.

Examples

if (requireNamespace("rstanarm", quietly=TRUE)) suppressWarnings({
    library(rstanarm)
    # fit a simple linear regression
    m = stan_glm(mpg ~ disp + cyl, data=mtcars,
        chains=1, iter=1000,
        control=list(adapt_delta=0.999), refresh=0)

    m = loo_conformal(m)
    # make predictive intervals
    predictive_interval(m)
})
#>                            5%      95%
#> Mazda RX4           17.272163 26.56781
#> Mazda RX4 Wag       17.338256 26.60573
#> Datsun 710          21.120373 31.16898
#> Hornet 4 Drive      15.282762 23.91217
#> Hornet Sportabout    9.692534 19.07198
#> Valiant             15.647767 25.36354
#> Duster 360           9.533651 19.32828
#> Merc 240D           19.976806 30.17477
#> Merc 230            20.575764 30.46546
#> Merc 280            17.254187 26.47874
#> Merc 280C           17.298873 26.36326
#> Merc 450SE          11.285085 20.94897
#> Merc 450SL          11.317589 20.84539
#> Merc 450SLC         11.781277 21.26017
#> Cadillac Fleetwood   6.645900 17.07670
#> Lincoln Continental  7.099779 17.41012
#> Chrysler Imperial    7.267956 17.59311
#> Fiat 128            21.332172 32.07739
#> Honda Civic         21.853715 31.42795
#> Toyota Corolla      21.957338 32.01999
#> Toyota Corona       20.910256 30.73708
#> Dodge Challenger    10.157514 20.32843
#> AMC Javelin         10.924511 20.17292
#> Camaro Z28           9.737651 19.37506
#> Pontiac Firebird     8.708599 18.61387
#> Fiat X1-9           21.940949 31.78036
#> Porsche 914-2       20.924055 30.82306
#> Lotus Europa        21.141494 31.12190
#> Ford Pantera L       9.611753 19.69293
#> Ferrari Dino        17.543158 26.80903
#> Maserati Bora       10.814080 20.51868
#> Volvo 142E          20.998102 30.89299